chore(): refactor types

+++ pkg/models/order.go
@@ -4,8 +4,8 @@ import "gorm.io/gorm"
 
 type Order struct {
 	gorm.Model
-	Id        int32 `json:"id" gorm:"primaryKey"`
-	Price     int32 `json:"price"`
-	ProductId int32 `json:"product_id"`
-	UserId    int32 `json:"user_id"`
+	Id        int64 `json:"id" gorm:"primaryKey"`
+	Price     int64 `json:"price"`
+	ProductId int64 `json:"product_id"`
+	UserId    int64 `json:"user_id"`
 }
